+++ AmethystAPI/src/minecraft/src/common/world/level/chunk/BlockVolume.hpp
@@ -1,6 +1,7 @@
 #pragma once
 #include <memory>
 #include <minecraft/src-deps/core/utility/buffer_span_mut.hpp>
+#include <minecraft/src/common/world/level/BlockPos.hpp>
 #include <minecraft/src/common/world/level/block/Block.hpp>
 #include <vector>
 
@@ -15,30 +16,15 @@ public:
     const Block* mInitBlock;
 
     // Reimplemented
-    BlockVolume(std::uint32_t width, std::uint32_t height, std::uint32_t depth, Block const* initBlock)
-        : mWidth(width), mHeight(height), mDepth(depth), mDimensionBottom(0), mInitBlock(initBlock)
-    {
-        mBlocks = buffer_span_mut<const Block*>(width * depth * height);
-
-        for (int i = 0; i < width * depth * height; i++) {
-            this->mBlocks.mBegin[i] = initBlock;
-        }
-    }
+    BlockVolume(std::uint32_t width, std::uint32_t height, std::uint32_t depth, Block const* initBlock);
 
     // Reimplemented
-    size_t index(uint32_t x, uint32_t y, uint32_t z)
-    {
-        return (this->mDepth * x + z) * this->mHeight + y;
-    }
+    size_t index(uint32_t x, uint32_t y, uint32_t z);
 
     // Utility functions not in minecraft
-    void set(const Block* block, uint32_t x, uint32_t y, uint32_t z)
-    {
-        this->mBlocks.mBegin[index(x, y, z)] = block;
-    }
+    void setBlock(const Block* block, uint32_t x, uint32_t y, uint32_t z);
+    void setBlock(const Block* block, BlockPos blockPos);
 
-    const Block* get(uint32_t x, uint32_t y, uint32_t z)
-    {
-        return this->mBlocks.mBegin[index(x, y, z)];
-    }
+    const Block* getBlock(uint32_t x, uint32_t y, uint32_t z);
+    const Block* getBlock(BlockPos blockPos);
 };
